mcp/zana-memory/pyproject.toml
[project]
name = "zana-memory"
version = "0.1.0"
description = "ZANA MCP Memory Server for ChromaDB"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "chromadb>=0.6.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
zana-memory = "server:main"

[tool.uv]
package = false
